* wxsmith_branch: updated wxthings from svn * wxsmith_branch: removed PCH usage for (small) wxDemo project * wxsmith_branch: compilation fixes related to wx 2.9.x

+++ src/plugins/contrib/wxSmithContribItems/wxthings/wxthings/include/wx/things/block.h
@@ -273,7 +273,7 @@ public :
             m_sort = source.GetSortType();
         }
 
-    inline int GetCount() const { return m_blocks.GetCount(); }
+    inline size_t GetCount() const { return m_blocks.GetCount(); }
 
     inline void Clear() { m_blocks.Clear(); }
 
@@ -357,7 +357,7 @@ public :
             m_sort = source.GetSortType();
         }
 
-    inline int GetCount() const { return m_blocks.GetCount(); }
+    inline size_t GetCount() const { return m_blocks.GetCount(); }
 
     inline void Clear() { m_blocks.Clear(); }
